Some code cleanup and optimising around setting hidden columns.
